+++ structs/CollegePlayerProgressions.go
@@ -1,33 +1,33 @@
 package structs
 
 type CollegePlayerProgressions struct {
-	Age             int
-	Stars           int
-	Overall         int
-	Stamina         int
-	Injury          int
-	FootballIQ      int
-	Speed           int
-	Carrying        int
-	Agility         int
-	Catching        int
-	RouteRunning    int
-	ZoneCoverage    int
-	ManCoverage     int
-	Strength        int
-	Tackle          int
-	PassBlock       int
-	RunBlock        int
-	PassRush        int
-	RunDefense      int
-	ThrowPower      int
-	ThrowAccuracy   int
-	KickAccuracy    int
-	KickPower       int
-	PuntAccuracy    int
-	PuntPower       int
-	Progression     int
-	Discipline      int
+	Age             int8
+	Stars           int8
+	Overall         int8
+	Stamina         int8
+	Injury          int8
+	FootballIQ      int8
+	Speed           int8
+	Carrying        int8
+	Agility         int8
+	Catching        int8
+	RouteRunning    int8
+	ZoneCoverage    int8
+	ManCoverage     int8
+	Strength        int8
+	Tackle          int8
+	PassBlock       int8
+	RunBlock        int8
+	PassRush        int8
+	RunDefense      int8
+	ThrowPower      int8
+	ThrowAccuracy   int8
+	KickAccuracy    int8
+	KickPower       int8
+	PuntAccuracy    int8
+	PuntPower       int8
+	Progression     int8
+	Discipline      int8
 	PotentialGrade  string
 	InjuryText      string
 	WeeksOfRecovery int
